Ajout de api_tocken pour restreindre l'acces

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Http\Request;
+use Illuminate\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -21,7 +22,7 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
 /**
  * Profil utilisateurs
  */
-Route::get('/profile/{userId}', 'ProfileController@show');
-Route::put('/profile/{userId}', 'ProfileController@store');
-Route::patch('/profile/{userId}', 'ProfileController@update');
-Route::delete('/profile/{userId}', 'ProfileController@destroy');
+Route::middleware('auth:api')->get('/profile/{userId}', 'ProfileController@show');
+Route::middleware('auth:api')->put('/profile/{userId}', 'ProfileController@store');
+Route::middleware('auth:api')->patch('/profile/{userId}', 'ProfileController@update');
+Route::middleware('auth:api')->delete('/profile/{userId}', 'ProfileController@destroy');
